prematch in PayloadBase
Adds an additional substring matching layer above the full payload match
functions as an optimization. If a particular payload type registers
prematch strings, one of these substrings must be found in the raw input
in order to attempt the full match prior to passing it off to the
parser.

If a payload type does not register any prematch strings, the full
payload match is always run.